requirements_cross_attention.txt
# Cross-Attention Plant Disease Classification Requirements

# Core ML Libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0

# Data Processing
numpy>=1.21.0
pandas>=1.4.0
Pillow>=9.0.0
scikit-learn>=1.1.0

# Configuration and Serialization
pyyaml>=6.0
dataclasses-json>=0.5.7

# Visualization and Plotting
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.9.0

# Progress and Logging
tqdm>=4.64.0

# Image Processing
opencv-python>=4.6.0
albumentations>=1.2.0

# Optional: For better performance
# nvidia-ml-py3>=7.352.0  # For GPU monitoring

# Development and Testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0